Made ExternMFD to Work

+++ Src/Orbiter/Psys.h
@@ -195,7 +195,7 @@ public:
 	void ActivatePlanetLabels(bool activate);
 
 	intptr_t FindFirst (int type, _finddata_t *fdata, char *path, char *fname);
-	int FindNext (intptr_t fh, _finddata_t *fdata, char *fname);
+	intptr_t FindNext (intptr_t fh, _finddata_t *fdata, char *fname);
 
 private:
 	char *name; // system's name
